wip: refactor cli command, add argument for single prompt generation
