+++ packages/@n8n/ai-workflow-builder/src/index.ts
@@ -1,4 +1,3 @@
 export * from './ai-workflow-builder.service';
-export * from './evaluator.service';
 export * from './types';
 export * from './workflow-state';
